revert 0cb5727f264472ba6bef4c7dc1518e854ea609a3 partially
all parts that used size_t casts for the int->void* issue
were removed again.

it seems the fix was not right, as motor crashes still on 64 bit sys
(maybe due to newer gcc versions which are more sensitive to UB).

+++ kkconsui/src/texteditor.cc
@@ -44,7 +44,6 @@
 */
 
 #include "texteditor.h"
-#include <stddef.h>
 
 #define CURLINE       (curfile ? (curfile->sy+curfile->y) : 0)
 #define CURCOL        (curfile ? (curfile->sx+curfile->x) : 0)
@@ -1944,11 +1943,11 @@ void texteditor::undorecordfree(void *p) {
 }
 
 int texteditor::findint(void *p1, void *p2) {
-    return *(size_t *) p1 != (size_t) p2;
+    return *(int *) p1 != (int) p2;
 }
 
 int texteditor::findhighline(void *p1, void *p2) {
-    return *(size_t *) p1 != ((highline *) p2)->line;
+    return *(int *) p1 != ((highline *) p2)->line;
 }
 
 void texteditor::shiftmarkedblock(int delta) {
